requirements.txt
openai-whisper==20230918
yt-dlp==2023.10.13
